Add files via upload
A simple code that shows the usage of fstream function and also the utilization of stucture in C++ that can be used to open and save files using code only. The two .txt file included is necessary so that the fstream function inside the code will execute. Also, kindly remember that the directory of both the .txt file and the source code must be the same.